添加设置导入视频信息长度设置为章节观看视频的长度

+++ databank/src/main/java/com.project.databank/web/vo/DataDatumVo.java
@@ -30,5 +30,5 @@ public class DataDatumVo implements Serializable {
 
     private int indexNum;
 
-    private long videoDuration;
+    private Integer videoDuration;
 }
